requirements.txt
flask==3.0.3
gunicorn==21.2.0
requests==2.31.0
pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
blockfrost-python==0.5.3
pdf2image==1.17.0
apscheduler==3.10.4
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pycardano==0.10.0
poppler-utils
firebase-admin==6.4.0
